Move utils file to separate files
