+++ apps/back-office/src/app/melding/[meldingId]/_utils/getLocationData.ts
@@ -1,0 +1,17 @@
+import type { MeldingOutput } from '@meldingen/api-client'
+
+import { getFullNLAddress } from '~/app/_utils/getFullNLAddress'
+
+export const getLocationData = (data: MeldingOutput, t: (key: string) => string) => {
+  const address = getFullNLAddress(data)
+
+  if (!address) return undefined
+
+  return [
+    {
+      description: address,
+      key: 'address',
+      term: t('detail.location.address'),
+    },
+  ]
+}
