+++ apps/back-office/src/app/melding/[meldingId]/_utils/index.ts
@@ -1,0 +1,5 @@
+export { getAdditionalQuestionsData } from './getAdditionalQuestionsData'
+export { getAttachmentsData } from './getAttachmentsData'
+export { getContactData } from './getContactData'
+export { getLocationData } from './getLocationData'
+export { getMeldingData } from './getMeldingData'
