--- apps/back-office/src/app/melding/[meldingId]/utils.test.tsx
@@ -1,322 +1,0 @@
-import { http, HttpResponse } from 'msw'
-
-import { getFullNLAddress } from '../../_utils/getFullNLAddress'
-import {
-  getAdditionalQuestionsData,
-  getAttachmentsData,
-  getContactData,
-  getLocationData,
-  getMeldingData,
-} from './utils'
-import { additionalTimeQuestion, additionalValueLabelQuestion, melding } from '~/mocks/data'
-import { additionalQuestions } from '~/mocks/data'
-import { ENDPOINTS } from '~/mocks/endpoints'
-import { server } from '~/mocks/node'
-
-const mockMeldingId = 88
-
-describe('getAdditionalQuestionsData', () => {
-  it('returns correct additional text question data', async () => {
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    const additionalQuestionsData = additionalQuestions.map((item) => ({
-      description: item.text,
-      key: item.question.id.toString(),
-      term: item.question.text,
-    }))
-
-    expect(result).toEqual({ data: additionalQuestionsData })
-  })
-
-  it('returns correct additional time question data', async () => {
-    server.use(http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ANSWERS, () => HttpResponse.json([additionalTimeQuestion])))
-
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    expect(result).toEqual({
-      data: [
-        {
-          description: additionalTimeQuestion.time,
-          key: additionalTimeQuestion.question.id.toString(),
-          term: additionalTimeQuestion.question.text,
-        },
-      ],
-    })
-  })
-
-  it('returns "Weet ik niet" for additional time questions when time is null', async () => {
-    const additionalTimeQuestionWithNullTime = {
-      ...additionalTimeQuestion,
-      time: null,
-    }
-
-    server.use(
-      http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ANSWERS, () =>
-        HttpResponse.json([additionalTimeQuestionWithNullTime]),
-      ),
-    )
-
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    expect(result).toEqual({
-      data: [
-        {
-          description: 'Weet ik niet',
-          key: additionalTimeQuestionWithNullTime.question.id.toString(),
-          term: additionalTimeQuestionWithNullTime.question.text,
-        },
-      ],
-    })
-  })
-
-  it('returns correct additional value_label question data', async () => {
-    server.use(
-      http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ANSWERS, () => HttpResponse.json([additionalValueLabelQuestion])),
-    )
-
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    expect(result).toEqual({
-      data: [
-        {
-          description: additionalValueLabelQuestion.values_and_labels
-            .map((valAndLabel) => valAndLabel.label)
-            .join(', '),
-          key: additionalValueLabelQuestion.question.id.toString(),
-          term: additionalValueLabelQuestion.question.text,
-        },
-      ],
-    })
-  })
-
-  it('returns an empty description for unsupported question types', async () => {
-    server.use(
-      http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ANSWERS, () =>
-        HttpResponse.json([{ question: { id: 3, text: 'Unsupported question type' }, type: 'unsupported_type' }]),
-      ),
-    )
-
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    expect(result).toEqual({
-      data: [{ description: '', key: '3', term: 'Unsupported question type' }],
-    })
-  })
-
-  it('returns an error message when error is returned', async () => {
-    server.use(
-      http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ANSWERS, () =>
-        HttpResponse.json({ detail: 'Error message' }, { status: 500 }),
-      ),
-    )
-    const result = await getAdditionalQuestionsData(mockMeldingId)
-
-    expect(result).toEqual({ error: 'Error message' })
-  })
-})
-
-describe('getContactData', () => {
-  it('returns correct contact data', () => {
-    const result = getContactData(melding, (key: string) => key)
-
-    expect(result).toEqual([
-      {
-        description: melding.email,
-        key: 'email',
-        term: 'detail.contact.email',
-      },
-      {
-        description: melding.phone,
-        key: 'phone',
-        term: 'detail.contact.phone',
-      },
-    ])
-  })
-
-  it('returns a fallback label when contact data does not exist', () => {
-    const meldingDataWithoutContact = {
-      ...melding,
-      email: null,
-      phone: null,
-    }
-
-    const result = getContactData(meldingDataWithoutContact, (key: string) => key)
-
-    expect(result).toEqual([
-      {
-        description: 'detail.contact.no-data',
-        key: 'email',
-        term: 'detail.contact.email',
-      },
-      {
-        description: 'detail.contact.no-data',
-        key: 'phone',
-        term: 'detail.contact.phone',
-      },
-    ])
-  })
-})
-
-describe('getLocationData', () => {
-  it('returns correct location data', () => {
-    const result = getLocationData(melding, (key: string) => key)
-
-    expect(result).toEqual([
-      {
-        description: getFullNLAddress(melding),
-        key: 'address',
-        term: 'detail.location.address',
-      },
-    ])
-  })
-
-  it('returns undefined when not all location data exists', () => {
-    const meldingDataWithoutPostalCode = {
-      ...melding,
-      postal_code: null,
-    }
-
-    const result = getLocationData(meldingDataWithoutPostalCode, (key: string) => key)
-
-    expect(result).toEqual(undefined)
-  })
-})
-
-describe('getMeldingData', () => {
-  it('returns correct melding summary', () => {
-    const result = getMeldingData(melding, (key: string) => key)
-
-    const { classification, created_at, id, labels, source } = melding
-
-    expect(result).toEqual([
-      {
-        description: new Date(created_at).toLocaleDateString('nl-NL'),
-        key: 'created_at',
-        term: 'detail.melding-data.created_at',
-      },
-      {
-        description: classification!.name,
-        key: 'classification',
-        term: 'detail.melding-data.classification.term',
-      },
-      {
-        description: `shared.state.${melding.state}`,
-        key: 'state',
-        link: {
-          href: `/melding/${id}/wijzig-status`,
-          label: 'detail.melding-data.state.link',
-        },
-        term: 'detail.melding-data.state.term',
-      },
-      {
-        description: `shared.urgency.${melding.urgency}`,
-        key: 'urgency',
-        link: {
-          href: `/melding/${id}/wijzig-urgentie`,
-          label: 'detail.melding-data.urgency.link',
-        },
-        term: 'detail.melding-data.urgency.term',
-      },
-      {
-        description: labels?.map((label) => label.name).join(', '),
-        key: 'labels',
-        link: {
-          href: `/melding/${id}/wijzig-labels`,
-          label: 'detail.melding-data.labels.link',
-        },
-        term: 'detail.melding-data.labels.term',
-      },
-      {
-        description: source?.name,
-        key: 'source',
-        term: 'detail.melding-data.source.term',
-      },
-    ])
-  })
-
-  it('returns fallback label when classification is null', () => {
-    const meldingDataWithoutClassification = {
-      ...melding,
-      classification: null,
-    }
-
-    const result = getMeldingData(meldingDataWithoutClassification, (key: string) => key)
-
-    const classificationEntry = result.find((item) => item.key === 'classification')
-
-    expect(classificationEntry?.description).toBe('detail.melding-data.classification.no-data')
-  })
-
-  it('returns fallback label when labels is empty', () => {
-    const meldingDataWithoutLabels = {
-      ...melding,
-      labels: [],
-    }
-
-    const result = getMeldingData(meldingDataWithoutLabels, (key: string) => key)
-
-    const labelsEntry = result.find((item) => item.key === 'labels')
-    expect(labelsEntry?.description).toBe('detail.melding-data.labels.no-data')
-  })
-
-  it('returns fallback label when source is null', () => {
-    const meldingDataWithoutSource = {
-      ...melding,
-      source: null,
-    }
-
-    const result = getMeldingData(meldingDataWithoutSource, (key: string) => key)
-
-    const sourceEntry = result.find((item) => item.key === 'source')
-    expect(sourceEntry?.description).toBe('detail.melding-data.source.no-data')
-  })
-})
-
-describe('getAttachmentsData', () => {
-  it('returns correct attachments data', async () => {
-    const result = await getAttachmentsData(mockMeldingId, (key: string) => key)
-
-    expect(result).toMatchObject({
-      files: [
-        {
-          blob: expect.any(Blob),
-          fileName: 'IMG_0815.jpg',
-        },
-      ],
-      key: 'attachments',
-      term: 'detail.attachments.title',
-    })
-  })
-
-  it('returns an error message when getMeldingByMeldingIdAttachments returns an error', async () => {
-    server.use(
-      http.get(ENDPOINTS.GET_MELDING_BY_MELDING_ID_ATTACHMENTS, () =>
-        HttpResponse.json({ detail: 'Error message' }, { status: 500 }),
-      ),
-    )
-
-    const result = await getAttachmentsData(mockMeldingId, (key: string) => key)
-
-    expect(result).toEqual({ error: 'Error message' })
-  })
-
-  it('returns an attachment with error message when getAttachmentById returns an error ', async () => {
-    server.use(
-      http.get(ENDPOINTS.GET_ATTACHMENT_BY_ID, () => HttpResponse.json({ detail: 'Error message' }, { status: 500 })),
-    )
-
-    const result = await getAttachmentsData(mockMeldingId, (key: string) => key)
-
-    expect(result).toMatchObject({
-      files: [
-        {
-          blob: null,
-          error: 'Error message',
-          fileName: 'IMG_0815.jpg',
-        },
-      ],
-      key: 'attachments',
-      term: 'detail.attachments.title',
-    })
-  })
-})
