--- apps/back-office/src/app/melding/[meldingId]/utils.ts
@@ -1,158 +1,0 @@
-import type {
-  GetMeldingByMeldingIdAnswersMelderResponses,
-  MeldingOutput,
-  ValueLabelObject,
-} from '~/app/_api-client/proxy'
-
-import { getFullNLAddress } from '../../_utils/getFullNLAddress'
-import {
-  getAttachmentById,
-  getMeldingByMeldingIdAnswers,
-  getMeldingByMeldingIdAttachments,
-} from '~/app/_api-client/proxy'
-import { handleApiError } from '~/app/_utils/handleApiError'
-
-const getDescription = (answer: GetMeldingByMeldingIdAnswersMelderResponses['200'][number]) => {
-  switch (answer.type) {
-    case 'text':
-      return answer.text
-    case 'time':
-      // If the Time value is null, it means the melder selected 'do not know' for a question about time.
-      // In that case, we want to show "Weet ik niet" instead of an empty string.
-      if (answer.time === null) return 'Weet ik niet'
-
-      return answer.time
-    case 'value_label':
-      return answer.values_and_labels.map((option: ValueLabelObject) => option.label).join(', ')
-    default:
-      return ''
-  }
-}
-
-export const getAdditionalQuestionsData = async (meldingId: number) => {
-  const { data, error } = await getMeldingByMeldingIdAnswers({
-    path: { melding_id: meldingId },
-  })
-
-  if (error) return { error: handleApiError(error) }
-
-  return {
-    data: data.map((answer) => ({
-      description: getDescription(answer),
-      key: String(answer.question.id),
-      term: answer.question.text,
-    })),
-  }
-}
-
-export const getContactData = (data: MeldingOutput, t: (key: string) => string) => {
-  const { email, phone } = data
-
-  return [
-    {
-      description: email ?? t('detail.contact.no-data'),
-      key: 'email',
-      term: t('detail.contact.email'),
-    },
-    {
-      description: phone ?? t('detail.contact.no-data'),
-      key: 'phone',
-      term: t('detail.contact.phone'),
-    },
-  ]
-}
-
-export const getMeldingData = (data: MeldingOutput, t: (key: string) => string) => {
-  const { classification, created_at, id, labels, source, state, urgency } = data
-
-  return [
-    {
-      description: new Date(created_at).toLocaleDateString('nl-NL'),
-      key: 'created_at',
-      term: t('detail.melding-data.created_at'),
-    },
-    {
-      description: classification ? classification.name : t('detail.melding-data.classification.no-data'),
-      key: 'classification',
-      term: t('detail.melding-data.classification.term'),
-    },
-    {
-      description: t(`shared.state.${state}`),
-      key: 'state',
-      link: {
-        href: `/melding/${id}/wijzig-status`,
-        label: t('detail.melding-data.state.link'),
-      },
-      term: t('detail.melding-data.state.term'),
-    },
-    {
-      description: t(`shared.urgency.${urgency}`),
-      key: 'urgency',
-      link: {
-        href: `/melding/${id}/wijzig-urgentie`,
-        label: t('detail.melding-data.urgency.link'),
-      },
-      term: t('detail.melding-data.urgency.term'),
-    },
-    {
-      description: labels?.length
-        ? labels.map((label) => label.name).join(', ')
-        : t('detail.melding-data.labels.no-data'),
-      key: 'labels',
-      link: {
-        href: `/melding/${id}/wijzig-labels`,
-        label: t('detail.melding-data.labels.link'),
-      },
-      term: t('detail.melding-data.labels.term'),
-    },
-    {
-      description: source ? source.name : t('detail.melding-data.source.no-data'),
-      key: 'source',
-      term: t('detail.melding-data.source.term'),
-    },
-  ]
-}
-
-export const getLocationData = (data: MeldingOutput, t: (key: string) => string) => {
-  const address = getFullNLAddress(data)
-
-  if (!address) return undefined
-
-  return [
-    {
-      description: address,
-      key: 'address',
-      term: t('detail.location.address'),
-    },
-  ]
-}
-
-export const getAttachmentsData = async (meldingId: number, t: (key: string) => string) => {
-  const { data, error } = await getMeldingByMeldingIdAttachments({
-    path: { melding_id: meldingId },
-  })
-
-  if (error) return { error: handleApiError(error) }
-
-  const attachments = await Promise.all(
-    data.map(async ({ id, original_filename }) => {
-      const { data, error } = await getAttachmentById({
-        path: { id },
-
-        query: { type: 'thumbnail' },
-      })
-
-      if (error) {
-        return { blob: null, error: handleApiError(error), fileName: original_filename }
-      }
-
-      // Returning blob instead of File since the File api is not available in Node.js
-      return {
-        blob: data as Blob,
-        fileName: original_filename,
-      }
-    }),
-  )
-
-  return { files: attachments, key: 'attachments', term: t('detail.attachments.title') }
-}
